WIP: migration class
